backend/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
packaging==23.2
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
celery==5.3.6
django-cors-headers==4.2.0
redis==5.0.4
daphne==4.0.0
channels==4.0.0
channels-redis==4.2.0
pywebpush==2.0.0
pillow==10.3.0
pylint==3.2.3
pylint-django==2.5.5
